requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-imagekit==4.0.2
gunicorn==20.0.4
pilkit==2.0
Pillow==7.1.1
psycopg2==2.8.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
